Move survey models to creation models module
