+++ src/rpft/parsers/creation/globalrowmodels.py
@@ -1,5 +1,5 @@
 from rpft.parsers.creation.datarowmodel import DataRowModel
-from rpft.parsers.creation.surveymodels import SurveyQuestionModel
+from rpft.parsers.creation.models import SurveyQuestionModel
 
 
 class SurveyQuestionRowModel(DataRowModel, SurveyQuestionModel):
